build(vite): include entry points for classes
Due to the way that the building process is carried out, each
`index.html` file should be declared individually for `vite` to be able
to find and compile the files. While also preserving the ability for
other students (which I got rid of, but for the purposes of only fixing
our own code) to include their own entry points without altering any
file structure nor requiring any changes to the links which work
perfectly in the `dev` environment.